Changed the AlphaLiveEngine receivedValue variable back to a float, as this broken the exp/log scaling of pressure. Can’t quite remember why I changed this to an int when I was adding the pressure latch feature - hopefully this hasn’t broken anything else.

+++ Source/Functionality Classes/AlphaLiveEngine.h
@@ -121,7 +121,7 @@ private:
     EliteControls *eliteControls;
     
     int recievedPad;
-    int recievedValue;
+    float recievedValue;
     float recievedVelocity;
     
     //midi output device
